% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_by_prev.R
\name{adjust_by_prev}
\alias{adjust_by_prev}
\title{Sample Size Calculator for Diagnostic Tests}
\usage{
adjust_by_prev(sample, prev, beta)
}
\arguments{
\item{prev}{(numeric): The expected prevalence of the target condition in the population.}

\item{beta}{(numeric): Type II error rate.}

\item{n_unadjusted}{(integer): The initial unadjusted sample size calculated without considering prevalence.}
}
\value{
Object of class \code{adjust_by_prev} containing:
\itemize{
\item method (character) - Adjustment method.
\item parameter (list) - Input parameters containing:
\itemize{
\item power (numeric) - Desired statistical power (1 - Type II error rate).
\item prevalence (numeric) - The expected prevalence of the target condition in the population.
}
\item sample (list) - Output results containing:
\itemize{
\item n_unadjusted (integer) - The initial unadjusted sample size calculated without considering prevalence.
\item n_adjusted (integer) - The total sample size should be recruited.
}
\item call (character) - The original function call.
}
}
\description{
Adjust sample size for prospective diagnostic accuracy studies, accounting for the prevalence of the condition.
}
\examples{
n_adj <- adjust_by_prev(sample=350, prev=0.8, beta=0.9)
print(n_adj)
}
