% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_for_one_fixed_test.R
\name{adjust_for_one_fixed_test}
\alias{adjust_for_one_fixed_test}
\title{Sample Size Calculator for Diagnostic Tests}
\usage{
adjust_for_one_fixed_test(n_known_test, n_unadjusted)
}
\arguments{
\item{n_known_test}{(integer): The known/predetermined sample size for one test, default as test 1.}

\item{n_unadjusted}{(integer): The desired total sample size for the study.}
}
\value{
Object of class \code{adjust_by_fixed_size} containing:
\itemize{
\item method (character) - Adjustment method.
\item parameter (list) - Input parameters containing:
\itemize{
\item n_known_test (integer) - The known/predetermined sample size for one test.
\item n_unadjusted (integer) - The desired total sample size for the study.
}
\item sample (list) - Output results containing:
\itemize{
\item n_unadj_known_test (integer) - The known/predetermined sample size for one test.
\item n_unadj_unknown_test (integer) - Unadjusted sample size for another test.
\item n_unadjusted (integer) - The desired total sample size for the study.
\item n_adj_known_test (integer) - The known/predetermined sample size for one test.
\item n_adj_unknown_test (integer) - The desired sample size for another test.
\item n_adjusted (integer) - The desired total sample size after adjustment.
}
\item call (character) - The original function call.
}
}
\description{
Adjust sample size for an unpaired study design comparing two diagnostic tests, when sample size for one test is fixed.
}
\examples{
n_adj <- adjust_for_one_fixed_test(n_known_test=52, n_unadjusted=77)
print(n_adj)
}
