% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_binorm_params.R
\name{get_binorm_params}
\alias{get_binorm_params}
\title{Sample Size Calculator for Diagnostic Tests}
\usage{
get_binorm_params(AUC, SD_Ratio)
}
\arguments{
\item{AUC}{(numeric): Areas under the ROC curve.}

\item{SD_Ratio}{(numeric): Ratio of standard deviations of the distributions of test results (without vs with condition).}
}
\value{
A list of class \code{binorm_params} containing:
\itemize{
\item a (numeric) - Standardized difference in means of the distributions of test results for patients with and without the condition, (μ_with - μ_without)/σ_with.
\item b (numeric) - Ratio of the standard deviations of the distributions of test results for patients without versus with the condition, σ_without/σ_with.
\item interpretation (list) - The interpretation of the ROC parameters (a, b).
\item call (character) - The original function call.
}
}
\description{
Estimate binormal model parameters (a, b) for Receiver Operating Characteristic (ROC) curve fitting.
}
\examples{
params <- get_binorm_params(AUC=0.8, SD_Ratio=1.2)
print(params)
}
