% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diag_sample.R
\name{get_diag_sample}
\alias{get_diag_sample}
\title{Sample Size Calculator for Diagnostic Tests}
\usage{
get_diag_sample(var_function, alpha = 0.05, beta = NULL, delta, test_type)
}
\arguments{
\item{var_function}{(numeric/vector). Variance function value(s) depending on test type:
\itemize{
\item For "one_diagnostic": a single numeric value.
\item For "two_diagnostic": a numeric vector of length 2 (V_null and V_alternative).
\item For "non_inferiority": a single numeric value.
\item For "equivalence": a single numeric value.
}}

\item{alpha}{(numeric): Significance level (default: 0.05).}

\item{beta}{(numeric): Type II error rate.}

\item{delta}{(numeric): Desired half-width of CI/Minimum detectable difference.}

\item{test_type}{(character). Type of test:
\itemize{
\item \code{"one_diagnostic"}: To evaluate the diagnostic accuracy of a single test.
\item \code{"two_diagnostic"}: To compare the diagnostic accuracy of two tests.
\item \code{"non_inferiority"}: To assess the non-inferiority of test 1 to test 2.
\item \code{"equivalence"}: To test the equivalence of test 1 and test 2.
}}
}
\value{
N (integer): Estimated sample size.
}
\description{
Calculates the required sample size for:
\enumerate{
\item estimating the accuracy of one diagnostic test;
\item compare the accuracies of two diagnostic tests;
\item assess non-inferiority or equivalence of two tests.
}
}
\examples{
N1 <- get_diag_sample(var_function=0.8, alpha=0.05, beta=0.2, delta=0.1, test_type="one_diagnostic")
print(N1)
N2 <- get_diag_sample(var_function=c(0.2, 0.3), alpha=0.05, beta=0.2, delta=0.1, test_type="two_diagnostic")
print(N2)
}
\references{
Zhou XH, Obuchowski NA, McClish DK. Statistical Methods in Diagnostic Medicine (2nd ed). Wiley; 2011.
}
