% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_for_threshold.R
\name{optimize_for_threshold}
\alias{optimize_for_threshold}
\title{Sample Size Calculator for Diagnostic Tests}
\usage{
optimize_for_threshold(
  spec_min,
  sens_at_spec_min,
  sens_min,
  alpha,
  beta,
  b = 1
)
}
\arguments{
\item{spec_min}{(numeric): The acceptable minimum specificity in order for the cutoff to be useful.}

\item{sens_at_spec_min}{(numeric): The conjectured sensitivity at the acceptable minimum specificity.}

\item{sens_min}{(numeric): The acceptable minimum sensitivity in order for the cutoff to be useful.}

\item{alpha}{(numeric): Significance level (default: 0.05).}

\item{beta}{(numeric): Type II error rate.}

\item{b}{(numeric): Ratio of the standard deviations of the distributions of test results for patients without versus with the condition (default: 1).}
}
\value{
Object of class \code{optimize_for_cutoff} containing:
\itemize{
\item method (character) - Adjustment method.
\item parameter (list) - Input parameters containing:
\itemize{
\item min_spec (numeric) - The acceptable minimum specificity in order for the cutoff to be useful.
\item sens_at_min_spec (numeric) - The conjectured sensitivity at the acceptable minimum specificity.
\item sens_min (numeric) - The acceptable minimum sensitivity in order for the cutoff to be useful.
\item alpha (numeric) - Significance level.
\item power (numeric) - Desired statistical power (1 - Type II error rate).
\item b (numeric) - Ratio of the standard deviations of the distributions of test results for patients without versus with the condition, σ_without/σ_with.
}
\item sample (list) - Output results containing:
\itemize{
\item n_total (integer) - The required total sample size.
}
\item call (character) - The original function call.
}
}
\description{
Estimate required sample size to identify a suitable cutoff value in diagnostic accuracy studies.
}
\examples{
n_total <- optimize_for_threshold(spec_min=0.9, sens_at_spec_min=0.6, sens_min=0.5, alpha=0.05, beta=0.2, b=1)
print(n_total)
}
