% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_binorm_roc.R
\name{plot_binorm_ROC}
\alias{plot_binorm_ROC}
\title{Sample Size Calculator for Diagnostic Tests}
\usage{
plot_binorm_ROC(a, b, n_sample = 1000, ci_level = 0.95, seed = 154)
}
\arguments{
\item{a}{(numeric): Standardized difference in means of the distributions of test results for patients with and without the condition, (μ_with - μ_without)/σ_with.}

\item{b}{(numeric): Ratio of the standard deviations of the distributions of test results for patients without versus with the condition, σ_without/σ_with.}

\item{n_sample}{(Integer): Half sample size for simulation (default: 1000).}

\item{ci_level}{(numeric):  Confidence level (default: 0.95).}

\item{seed}{(Integer): Random seed (default: 154).}
}
\value{
A ggplot object showing the ROC curve with the following elements:
\itemize{
\item plot (ggplot object) - The ROC curve under binormal model.
\item n (integar) - The total sample size.
\item data (list) - Randomly generated simulation data (scores, labels).
\item call (character) - The original function call.
}
}
\description{
Fit a Receiver Operating Characteristic (ROC) curve by assuming normal distribution normally distributed test results for both cases (with condition) and controls (without condition) groups.
}
\examples{
p <- plot_binorm_roc(a=1.19, b=1, n_sample=2000, ci=0.95, seed=1)
print(p)
}
