% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.adjust_by_prev.R
\name{print.adjust_by_prev}
\alias{print.adjust_by_prev}
\title{Print method}
\usage{
\method{print}{adjust_by_prev}(x)
}
\arguments{
\item{x}{A \code{adjust_by_prev} object returned by \code{\link[=adjust_by_disease_prevalence]{adjust_by_disease_prevalence()}}.}
}
\value{
Invisibly returns for \code{adjust_by_prev} object.
}
\description{
Display a human-readable summary of adjust_by_prev objects containing sample size before and after adjustment.
}
\examples{
n_adj <- adjust_by_disease_prevalence(n_unadjusted=350, prev=0.8, beta=0.9)
print(n_adj)
n_adj  # same as print(n_adj)
}
