% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.adjust_by_fixed_size.R
\name{print.adjust_fixed_size}
\alias{print.adjust_fixed_size}
\title{Print method}
\usage{
\method{print}{adjust_fixed_size}(x)
}
\arguments{
\item{x}{A \code{adjust_by_fixed_size} object returned by \code{\link[=adjust_for_one_fixed_test]{adjust_for_one_fixed_test()}}.}
}
\value{
Invisibly returns for \code{adjust_by_fixed_size} object.
}
\description{
Display a human-readable summary of adjust_by_fixed_size objects containing sample size before and after adjustment.
}
\examples{
n_adj <- adjust_for_one_fixed_test(n_known_test=52, n_unadjusted=77)
print(n_adj)
n_adj  # same as print(n_adj)
}
