% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.adjust_for_cluster_data.R
\name{print.adjust_for_cluster_data}
\alias{print.adjust_for_cluster_data}
\title{Print method}
\usage{
\method{print}{adjust_for_cluster_data}(x)
}
\arguments{
\item{x}{A \code{adjust_by_cluster} object returned by \code{\link[=adjust_for_cluster_data]{adjust_for_cluster_data()}}.}
}
\value{
Invisibly returns for \code{adjust_by_cluster} object.
}
\description{
Display a human-readable summary of adjust_by_cluster objects containing sample size before and after adjustment.
}
\examples{
n_adj <- adjust_for_cluster_data(n1_unadj=46, avg_obs_per_patient=1.5, intra_corr=0.5)
print(n_adj)
n_adj  # same as print(n_adj)
}
