% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.optimize_for_cutoff.R
\name{print.optimize_for_cutoff}
\alias{print.optimize_for_cutoff}
\title{Print method}
\usage{
\method{print}{optimize_for_cutoff}(x)
}
\arguments{
\item{x}{A \code{optimize_for_cutoff} object returned by \code{\link[=optimize_for_threshold]{optimize_for_threshold()}}.}
}
\value{
Invisibly returns for \code{optimize_for_cutoff} object.
}
\description{
Display a human-readable summary of optimize_for_cutoff objects containing the required sample size for deciding a suitable cut-off.
}
\examples{
n <- optimize_for_threshold(spec_min=0.9, sens_at_spec_min=0.6, sens_min=0.5, alpha=0.05, beta=0.2, b=1)
print(n)
n # same as print(n_adj)
}
