% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.plot_binorm_roc.R
\name{print.plot_binorm_roc}
\alias{print.plot_binorm_roc}
\title{Print method}
\usage{
\method{print}{plot_binorm_roc}(x)
}
\arguments{
\item{x}{A \code{plot_binorm_roc} object returned by \code{\link[=plot_binorm_ROC]{plot_binorm_ROC()}}.}
}
\value{
Invisibly returns for \code{plot_binorm_ROC} object.
}
\description{
Displays a structured summary of \code{plot_binorm_ROC} objects, including the function call, sample size, and simulation data, followed by the ROC curve plot.
}
\examples{
plot <- plot_binorm_ROC(a=1.19, b=1, n_sample=2000, ci=0.95, seed=1)
print(plot)
plot  # same as print(plot)
}
