% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_compare_partial_roc.R
\name{sample_compare_partial_roc}
\alias{sample_compare_partial_roc}
\title{Sample Size Calculator for Diagnostic Tests}
\usage{
sample_compare_partial_roc(
  A_null,
  b_null = 1,
  A1_alter,
  b1_alter = 1,
  A2_alter,
  b2_alter = 1,
  e1,
  e2,
  rN = 0.5,
  rD = 0.5,
  alpha = 0.05,
  beta,
  R = 1,
  paired = FALSE
)
}
\arguments{
\item{A_null}{(numeric): AUC under the null hypothesis.}

\item{b_null}{(numeric): Ratio of the standard deviations of the distributions of test results for patients without versus with the condition, σ_without/σ_with (default: 1).}

\item{A1_alter}{(numeric): AUC of test 1 under the alternative hypothesis.}

\item{b1_alter}{(numeric): Ratio of the standard deviations of the distributions of results for test 1 for patients without versus with the condition, σ1_without/σ1_with (default: 1).}

\item{A2_alter}{(numeric): AUC of test 2 under the alternative hypothesis.}

\item{b2_alter}{(numeric): Ratio of the standard deviations of the distributions of results for test 2 for patients without versus with the condition, σ2_without/σ2_with (default: 1).}

\item{e1}{(numeric): Lower bound of the false positive rate (FPR) range.}

\item{e2}{(numeric): Upper bound of the false positive rate (FPR) range.}

\item{rN}{(numeric): The correlation of the underlying bivariate binormal distribution for patients without the condition (default: 0.5).}

\item{rD}{(numeric): The correlation of the underlying bivariate binormal distribution for patients with the condition (default: 0.5).}

\item{alpha}{(numeric): Significance level (default: 0.05).}

\item{beta}{(numeric): Type II error rate.}

\item{R}{(numeric): Ratio of patients with and without the condition (default: 1).}

\item{paired}{(logical). Whether the design is paired (\code{TRUE}) or unpaired (\code{FALSE}).}
}
\value{
An object of class "diag_sample_size_html" containing:
\itemize{
\item \code{sample_size} (list): Required sample size.
\item \code{parameters} (list): Input parameters.
\item \code{html_report} (html): Sample size calculation report.
}
}
\description{
Calculate the required sample size for comparing partial Area Under the ROC Curve (pAUC) between two diagnostic tests.
}
\examples{
n <- sample_compare_partial_roc(A_null= 0.8, b_null=1, A1_alter=0.8, b1_alter=1, A2_alter=0.9, b2_alter=1, e1=0, e2=0.2, rN=0, rD=0, alpha = 0.05, beta=0.2, paired=FALSE)
print(n)
}
