% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_compare_ppvnpv.R
\name{sample_compare_ppvnpv}
\alias{sample_compare_ppvnpv}
\title{Sample Size Calculator for Diagnostic Tests}
\usage{
sample_compare_ppvnpv(
  metric = "PPV",
  theta2,
  p,
  alpha = 0.05,
  beta,
  gamma,
  delta,
  paired = TRUE,
  alternative = "two.sided"
)
}
\arguments{
\item{theta2}{Numeric. NPV or PPV of test 2 (0 < theta2 < 1).}

\item{p}{Numeric vector (length = 8). Probability components required for calculation.
\describe{
For patients without the condition：
\item{p\link{1}}{the proportion of patients testing positive on both tests}
\item{p\link{2}}{the proportion testing positive on test 1 and negative on test 2}
\item{p\link{3}}{the proportion testing positive on test 2 and negative on test 1}
\item{p\link{4}}{the proportion testing negative on both tests}
For patients with the condition：
\item{p\link{5}}{the proportion of patients testing positive on both tests}
\item{p\link{6}}{the proportion testing positive on test 1 and negative on test 2}
\item{p\link{7}}{the proportion testing positive on test 2 and negative on test 1}
\item{p\link{8}}{the proportion testing negative on both tests}
}}

\item{alpha}{Numeric. Significance level (0 < alpha < 1).}

\item{beta}{Numeric. Type II error rate (0 < beta < 1).}

\item{gamma}{Numeric. A specific value of interest for rPPV under the alternative hypothesis (gamma > 0).}

\item{delta}{Numeric. The null hypothesis is rPPV = delta. (In most cases, let delta = 1).}

\item{paired}{Logical. Whether the design is paired (\code{TRUE}) or unpaired (\code{FALSE}).}

\item{alternative}{Character. \code{"less"}, \code{"greater"}, or \code{"two.sided"}.}

\item{choice}{Character. \code{"PPV"} for positive predictive value, \code{"NPV"} for negative predictive value.}
}
\value{
An object of class "diag_sample_size_html" containing:
\itemize{
\item \code{sample_size} (list): Required sample size.
\item \code{parameters} (list): Input parameters.
\item \code{html_report} (html): Sample size calculation report.
}
}
\description{
Calculate the required sample size for comparing the positive predictive value (PPV) or negative predictive value (NPV) between two diagnostic tests.
}
\examples{
n1 <- sample_compare_ppvnpv(theta2=0.8, p=c(0.4,0,0.52,0.44,0.7,0,0.1,0.2), alpha=0.05, beta=0.2, delta=1, gamma=1.1, paired=TRUE, metric="PPV", alternative="two.sided")
print(n1)
n2 <- sample_compare_ppvnpv(theta2=0.8, p=c(0.4,0,0.52,0.44,0.7,0,0.1,0.2), alpha=0.10, beta=0.2, delta=1, gamma=1.37, paired=TRUE, metric="NPV", alternative="two.sided")
print(n2)
}
