% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_compare_restricted_sesp.R
\name{sample_compare_restricted_sesp}
\alias{sample_compare_restricted_sesp}
\title{Sample Size Calculator for Diagnostic Tests}
\usage{
sample_compare_restricted_sesp(
  metric,
  A_null,
  b_null = 1,
  A1_alter,
  b1_alter = 1,
  A2_alter,
  b2_alter = 1,
  e,
  rN = 0.5,
  rD = 0.5,
  alpha = 0.05,
  beta,
  paired = TRUE,
  R = 1
)
}
\arguments{
\item{metric}{(character): Performance metric to be evaluated, must be either "sens" or "sepc":
\itemize{
\item \code{"sens"}: Sensitivity at the fixed FPR.
\item \code{"spec"}: Specificity at the fixed FNR.
}}

\item{A_null}{(numeric): AUC under the null hypothesis.}

\item{b_null}{(numeric): Ratio of the standard deviations of the distributions of test results for patients without versus with the condition, σ_without/σ_with (default: 1).}

\item{A1_alter}{(numeric): AUC of test 1 under the alternative hypothesis.}

\item{b1_alter}{(numeric): Ratio of standard deviations of test 1 under the alternative hypothesis.}

\item{A2_alter}{(numeric): AUC of test 2 under the alternative hypothesis.}

\item{b2_alter}{(numeric): Ratio of standard deviations of test 2 under the alternative hypothesis.}

\item{e}{(numeric): For "sens" provide false positive rate value, and for "spec" provide false negative rate value.}

\item{rN}{(numeric): The correlation of the underlying bivariate binormal distribution for patients without the condition (default: 0.5).}

\item{rD}{(numeric): The correlation of the underlying bivariate binormal distribution for patients with the condition (default: 0.5).}

\item{alpha}{(numeric): Significance level (default: 0.05).}

\item{beta}{(numeric): Type II error rate.}

\item{paired}{(logical). Whether the design is paired (\code{TRUE}) or unpaired (\code{FALSE}).}

\item{R}{(numeric): Ratio of patients with and without the condition (default: 1).}

\item{dist}{(Character): The assumption for choosing the variance function (default: "binorm"):
\itemize{
\item \code{"binorm"}: Assume that the unobserved, underlying test results follow a binormal distribution, but the observed test results, either continuous or ordinal, do not necessarily have a binormal distribution.
\item \code{"obs_binorm"}: Assume that the observed test results are on a truly continuous scale and they follow a binormal distribution (or can be transformed to a binormal distribution).
}}
}
\value{
An object of class "diag_sample_size_html" containing:
\itemize{
\item \code{sample_size} (list): Required sample size.
\item \code{parameters} (list): Input parameters.
\item \code{html_report} (html): Sample size calculation report.
}
}
\description{
Calculate the required sample size for comparing sensitivity at a fixed false positive rate (FPR) or specificity at a fixed false negative rate (FNR) of two diagnostic tests.
}
\examples{
n <- sample_compare_restricted_sesp(metric="sens", A_null=0.8, b_null=1, A1_alter=0.8, b1_alter=1, A2_alter=0.9, b2_alter=1, e=0.1, rN=0.5, rD=0.5, alpha=0.05, beta=0.2, paired=TRUE, R=1, dist ="binorm")
print(n)
}
