% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_equivalence_auc.R
\name{sample_equivalence_auc}
\alias{sample_equivalence_auc}
\title{Sample Size Calculator for Diagnostic Tests}
\usage{
sample_equivalence_auc(
  As,
  bs = 1,
  Ae,
  be = 1,
  r = NULL,
  rN = NULL,
  rD = NULL,
  delta,
  alpha = 0.05,
  beta,
  R = 1,
  paired = TRUE,
  dist = "any"
)
}
\arguments{
\item{As}{(numeric): The expected sensitivity (or specificity) of an existing test (standard test).}

\item{Ae}{(numeric): The expected sensitivity (or specificity) of a new test (experimental test).}

\item{delta}{(numeric):  The prespecified upper clinical limits for equivalence.}

\item{alpha}{(numeric): Significance level (default: 0.05).}

\item{beta}{(numeric): Type II error rate.}

\item{paired}{(logical): Whether the design is paired (\code{TRUE}) or unpaired (\code{FALSE}).}

\item{dist}{(character): The assumption for choosing the variance function (default: "any"):
\itemize{
\item \code{"any"}: Applicable for all distributions.
\item \code{"binorm"}: Assume that the test results have an underlying bivariate binormal distribution.
}}

\item{prob_t1pos_give_t2pos}{(numeric): Conditional probability of the existing test being positive when the new test being positive.}
}
\value{
An object of class "diag_sample_size_html" containing:
\itemize{
\item \code{sample_size} (list): Required sample size.
\item \code{parameters} (list): Input parameters.
\item \code{html_report} (html): Sample size calculation report.
}
}
\description{
Calculate the required sample size for testing whether the Area Under the ROC Curve (AUC) of a new test is equivalent to an existing test.
}
\examples{
n <- sample_equivalence_auc(As=0.9, bs=1, Ae=0.88, be=1, rN=0.5, rD=0.5, delta=0.05, alpha=0.05, beta=0.2, R=1, paired=TRUE, dist="binorm")
n <- sample_equivalence_auc(As=0.9, bs=1, Ae=0.90, be=1, rN=0.5, rD=0.5, delta=0.05, alpha=0.05, beta=0.2, R=1, paired=TRUE, dist="binorm")
n <- sample_equivalence_auc(As=0.9, bs=1, Ae=0.92, be=1, rN=0.5, rD=0.5, delta=0.05, alpha=0.05, beta=0.2, R=1, paired=TRUE, dist="binorm")
print(n)
}
