% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_equivalence_sesp.R
\name{sample_equivalence_sesp}
\alias{sample_equivalence_sesp}
\title{Sample Size Calculator for Diagnostic Tests}
\usage{
sample_equivalence_sesp(
  theta_S,
  theta_E,
  prob_t1pos_give_t2pos = NULL,
  delta,
  alpha = 0.05,
  beta,
  R = 1,
  paired = TRUE
)
}
\arguments{
\item{theta_S}{(numeric): The expected sensitivity (or specificity) of an existing test (standard test).}

\item{theta_E}{(numeric): The expected sensitivity (or specificity) of a new test (experimental test).}

\item{prob_t1pos_give_t2pos}{(numeric): Conditional probability of the existing test being positive when the new test being positive.}

\item{delta}{(numeric): The prespecified upper clinical limits for equivalence.}

\item{alpha}{(numeric): Significance level (default: 0.05).}

\item{beta}{(numeric): Type II error rate.}

\item{paired}{(character): Whether the design is paired (\code{TRUE}) or unpaired (\code{FALSE}).}
}
\value{
An object of class "diag_sample_size_html" containing:
\itemize{
\item \code{sample_size} (list): Required sample size.
\item \code{parameters} (list): Input parameters.
\item \code{html_report} (html): Sample size calculation report.
}
}
\description{
Calculate the required sample size for testing whether the sensitivity/specificity of a new test is equivalent to an existing test.
}
\examples{
n <- sample_equivalence_sesp(theta_S=0.9, theta_E=0.88, prob_t1pos_give_t2pos=0, delta=0.05, alpha=0.05, beta=0.2, R=1, paired=FALSE)
print(n)
}
