% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_estimate_full_ROC.R
\name{sample_estimate_full_ROC}
\alias{sample_estimate_full_ROC}
\title{Sample Size Calculator for Diagnostic Tests}
\usage{
sample_estimate_full_ROC(
  A,
  b = 1,
  alpha = 0.05,
  beta = NULL,
  L,
  R = 1,
  dist = "any"
)
}
\arguments{
\item{A}{(numeric): Expected area under the ROC curve.}

\item{b}{(numeric): Ratio of the standard deviations of the distributions of test results for patients without versus with the condition, σ_without/σ_with (default: 1).}

\item{alpha}{(Numeric): Significance level (default: 0.05).}

\item{beta}{(Numeric): Type II error rate (default: NULL).}

\item{L}{(Numeric): The desired length of one-half of the (1-α)×100\% confidence interval for AUC.}

\item{R}{(Numeric): Ratio of patients with and without the condition (default: 1).}

\item{dist}{(Character): The assumption for choosing the variance function (default: "any"):
\itemize{
\item \code{"any"}: Applicable for all distributions.
\item \code{"exp"}: Assume that the unobserved, underlying test results follow an exponential distribution, but the observed test results, either continuous or ordinal, do not necessarily have an exponential distribution
\item \code{"binorm"}: Assume that the unobserved, underlying test results follow a binormal distribution, but the observed test results, either continuous or ordinal, do not necessarily have a binormal distribution
\item \code{"obs_binorm"}: Assume that the observed test results are on a truly continuous scale and follow a binormal distribution or can be transformed to a binormal distribution.
}}
}
\value{
An object of class "diag_sample_size_html" containing:
\itemize{
\item \code{sample_size} (list): Required sample size.
\item \code{parameters} (list): Input parameters.
\item \code{html_report} (html): Sample size calculation report.
}
}
\description{
Calculate the required sample size for estimating the Area Under the ROC Curve (AUC) of a diagnostic test.
}
\examples{
lapply(X=c("any", "exp", "binorm", "obs_binorm"), FUN=function(dist) sample_estimate_full_ROC(A=0.8, alpha=0.05, beta=0.2, L=0.1, R=1, dist=dist))
}
