% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_estimate_pauc.R
\name{sample_estimate_pauc}
\alias{sample_estimate_pauc}
\title{Sample Size Calculator for Diagnostic Tests}
\usage{
sample_estimate_pauc(
  A,
  b = 1,
  e1,
  e2,
  alpha = 0.05,
  beta = NULL,
  L,
  R = 1,
  dist = "binorm"
)
}
\arguments{
\item{A}{(numeric): Expected area under the ROC curve.}

\item{b}{(numeric): Ratio of the standard deviations of the distributions of test results for patients without versus with the condition, σ_without/σ_with (default: 1).}

\item{e1}{(numeric): Lower bound of the false positive rate (FPR) range.}

\item{e2}{(numeric): Upper bound of the false positive rate (FPR) range.}

\item{alpha}{(numeric): Significance level (default: 0.05).}

\item{beta}{(numeric): Type II error rate (default: NULL).}

\item{L}{(numeric): The desired length of one-half of the (1-α)×100\% confidence interval for partial AUC.}

\item{R}{(Numeric): Ratio of patients without to with the condition (default: 1).}

\item{dist}{(character): The assumption for choosing the variance function (default: "binorm"):
\itemize{
\item \code{"binorm"}: Assume that the unobserved, underlying test results follow a binormal distribution, but the observed test results, either continuous or ordinal, do not necessarily have a binormal distribution.
\item \code{"obs_binorm"}: The observed test results follow a binormal distribution.
}}
}
\value{
An object of class "diag_sample_size_html" containing:
\itemize{
\item \code{sample_size} (list): Required sample size.
\item \code{parameters} (list): Input parameters.
\item \code{html_report} (html): Sample size calculation report.
}
}
\description{
Calculate the required sample size for estimating the partial Area Under the ROC Curve (pAUC) of a diagnostic test.
}
\examples{
n <- sample_estimate_pauc(A=0.8, e1=0, e2=0.1, b=1, alpha=0.05, beta=NULL, L=0.1, R=1, dist="binorm")
print(n)
}
