% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_estimate_restricted_sesp.R
\name{sample_estimate_restricted_sesp}
\alias{sample_estimate_restricted_sesp}
\title{Sample Size Calculator for Diagnostic Tests}
\usage{
sample_estimate_restricted_sesp(
  metric,
  A,
  b = 1,
  e,
  alpha = 0.05,
  beta = NULL,
  L,
  R = 1,
  dist = "binorm"
)
}
\arguments{
\item{metric}{(character): Performance metric to be evaluated, must be either "sens" or "sepc":
\itemize{
\item \code{"sens"}: Sensitivity at the fixed FPR.
\item \code{"spec"}: Specificity at the fixed FNR.
}}

\item{A}{(numeric): Expected area under the ROC curve.}

\item{b}{(numeric): Ratio of the standard deviations of the distributions of test results for patients without versus with the condition, σ_without/σ_with (default: 1).}

\item{e}{(Numeric): For "sens" provide false positive rate value, and for "spec" provide false negative rate value.}

\item{alpha}{(Numeric): Significance level (default: 0.05).}

\item{beta}{(Numeric): Type II error rate (default: NULL).}

\item{L}{(Numeric): The desired length of one-half of the (1-α)×100\% confidence interval for sensitivity/specificity.}

\item{R}{(Numeric): Ratio of patients with and without the condition (default: 1).}

\item{dist}{(Character): The assumption for choosing the variance function (default: "binorm"):
\itemize{
\item \code{"binorm"}: Assume that the unobserved, underlying test results follow a binormal distribution, but the observed test results, either continuous or ordinal, do not necessarily have a binormal distribution.
\item \code{"obs_binorm"}: Assume that the observed test results are on a truly continuous scale and they follow a binormal distribution (or can be transformed to a binormal distribution).
}}
}
\value{
An object of class "diag_sample_size_html" containing:
\itemize{
\item \code{sample_size} (list): Required sample size.
\item \code{parameters} (list): Input parameters.
\item \code{html_report} (html): Sample size calculation report.
}
}
\description{
Calculate the required sample size for estimating sensitivity at a fixed false positive rate (FPR) or specificity at a fixed false negative rate (FNR) of a diagnostic test.
}
\examples{
n1 <- sample_estimate_restricted_sesp(metric="sens", A=0.8, b=1, e=0.1, alpha=0.05, beta=NULL, L=0.1, R=1, dist="binorm")
print(n1)
n2 <- sample_estimate_restricted_sesp(metric="spec", A=0.8, b=1, e=0.1, alpha=0.05, beta=0.2, L=0.1, R=1, dist="binorm")
print(n2)
}
