% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_estimate_sesp.R
\name{sample_estimate_sesp}
\alias{sample_estimate_sesp}
\title{Sample Size Calculator for Diagnostic Tests}
\usage{
sample_estimate_sesp(theta, alpha = 0.05, beta = NULL, L)
}
\arguments{
\item{theta}{(numeric). The conjectured sensitivity or specificity.}

\item{alpha}{(numeric). Significance level (default: 0.05).}

\item{beta}{(numeric). Type II error rate (default: NULL).}

\item{L}{(numeric). The desired length of one-half of the (1-α)×100\% confidence interval for sensitivity/specificity.}
}
\value{
An object of class "diag_sample_size_html" containing:
\itemize{
\item \code{sample_size}(list): Required sample size.
\item \code{parameters}(list): Input parameters.
\item \code{html_report}(html): Sample size calculation report.
}
}
\description{
Calculate the required sample size for estimating sensitivity or specificity of a single diagnostic test.
}
\examples{
n1 <- sample_estimate_sesp(theta=0.8, alpha=0.05, beta=0.2, L=0.05) #502
print(n1)
n2 <- sample_estimate_sesp(theta=0.8, alpha=0.05, L=0.05)
print(n2)
}
