% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_hypotest_auc.R
\name{sample_hypotest_auc}
\alias{sample_hypotest_auc}
\title{Sample Size Calculation for Fixed AUC Comparison}
\usage{
sample_hypotest_auc(
  A_null,
  A_alter,
  b = 1,
  alpha = 0.05,
  beta,
  alternative = "two.sided",
  R = 1
)
}
\arguments{
\item{A_null}{Numeric, assumed null hypothesis AUC (0.5 < A_null < 1).}

\item{A_alter}{Numeric, alternative hypothesis AUC (0.5 < A_alter < 1).}

\item{b}{Numeric, standard deviation ratio (default is 1, must be > 0).}

\item{alpha}{Numeric, significance level (0 < alpha < 1).}

\item{beta}{Numeric, Type II error (0 < beta < 1, power = 1 - beta).}

\item{alternative}{Character, type of test, must be one of "less", "greater", or "two.sided" (default: "two.sided").}

\item{R}{Numeric, ratio of non-diseased to diseased subjects (default is 1, must be > 0).}
}
\value{
An object of class "diag_sample_size_html" containing:
\itemize{
\item \code{sample_size} (list): Required sample size.
\item \code{parameters} (list): Input parameters.
\item \code{html_report} (html): Sample size calculation report.
}
}
\description{
Computes the required sample size for testing whether the AUC is equal to a specific value.
}
\details{
Sample Size Calculator for Diagnostic Tests

Calculate the required sample size for testing the hypothesis that the Area Under the ROC Curve is equal to a particular value.
}
\examples{
n <- sample_hypotest_auc(A_null=0.5, A_alter=0.8, alpha=0.05, beta=0.2, alternative="two.sided", R=1)
print(n)
}
