% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_hypotest_sesp.R
\name{sample_hypotest_sesp}
\alias{sample_hypotest_sesp}
\title{Sample Size Calculator for Diagnostic Tests}
\usage{
sample_hypotest_sesp(
  theta_null,
  theta_alter,
  alpha = 0.05,
  beta = NULL,
  alternative = "two.sided",
  R = 1
)
}
\arguments{
\item{alpha}{(numeric): Significance level (default: 0.05).}

\item{beta}{(numeric): Type II error rate (default: NULL).}

\item{alternative}{(character): Type of test, must be one of "less", "greater", or "two.sided" (default: "two.sided").}

\item{R}{(numeric): Ratio of patients without to with the condition (default: 1).}

\item{theta}{(numeric): The conjectured sensitivity or specificity.}
}
\value{
An object of class "diag_sample_size_html" containing:
\itemize{
\item \code{sample_size} (list): Required sample size.
\item \code{parameters} (list): Input parameters.
\item \code{html_report} (html): Sample size calculation report.
}
}
\description{
Calculate the required sample size for testing the hypothesis that sensitivity or specificity of a new test is equal to a particular value.
}
\examples{
n1 <- sample_hypotest_sesp(theta_null=0.5, theta_alter=0.8, alpha=0.05, beta=0.2, alternative="two.sided", R=1) #20
print(n1)
}
