% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_noninferior_auc.R
\name{sample_non_inferior_auc}
\alias{sample_non_inferior_auc}
\title{Sample Size Calculator for Diagnostic Tests}
\usage{
sample_non_inferior_auc(
  AUC_S = NULL,
  b_S = 1,
  AUC_E = NULL,
  b_E = 1,
  r = 0.5,
  rN = 0.5,
  rD = 0.5,
  delta,
  alpha = 0.05,
  beta,
  R = 1,
  paired = FALSE,
  dist = "any"
)
}
\arguments{
\item{AUC_S}{(numeric): The true area under the ROC curve (AUC) of an established standard test (test S).}

\item{b_S}{(numeric): Ratio of the standard deviations of the distributions of test results for test S for patients without versus with the condition, ﾏピ_without/ﾏピ_with (default: 1).}

\item{AUC_E}{(numeric): The true area under the ROC curve (AUC) of a new experimental test (test E).}

\item{b_E}{(numeric): Ratio of the standard deviations of the distributions of results for test E for patients without versus with the condition, ﾏテ_without/ﾏテ_with (default: 1).}

\item{r}{(numeric): The correlation between the tests because of the paired design.}

\item{rN}{(numeric): The correlation of the underlying bivariate binormal distribution for patients without the condition (default: 0.5).}

\item{rD}{(numeric): The correlation of the underlying bivariate binormal distribution for patients with the condition (default: 0.5).}

\item{delta}{(numeric): The smallest difference in accuracy which would not be considered non-inferior.}

\item{alpha}{(numeric): Significance level (default: 0.05).}

\item{beta}{(numeric): Type II error rate.}

\item{R}{(numeric): Ratio of patients with and without the condition.}

\item{paired}{(logical): Whether the design is paired (\code{TRUE}) or unpaired (\code{FALSE}).}

\item{dist}{(character): The assumption for choosing the variance function (default: "any").
\itemize{
\item \code{"any"}: Applicable for tests with any underlying distributions.
\item \code{"binorm"}: Assume that the test results have an underlying bivariate binormal distribution.
}}
}
\value{
An object of class "diag_sample_size_html" containing:
\itemize{
\item \code{sample_size} (list): Required sample size.
\item \code{parameters} (list): Input parameters.
\item \code{html_report} (html): Sample size calculation report.
}
}
\description{
Calculate the required sample size for testing whether the area under the ROC curve (AUC) of a new experimental test (test E) is non-inferior to an established standard test (test S).
}
\examples{
n <- sample_non_inferior_auc(AUC_S=0.9, AUC_E=0.88, b_S=1, b_E=1, r=0.5, rN=0.5, rD=0.5, delta=0.05, alpha=0.05, beta=0.2, R=1, paired=TRUE, dist="binorm")
print(n)
}
