% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_noninferior_sesp.R
\name{sample_non_inferior_sesp}
\alias{sample_non_inferior_sesp}
\title{Sample Size Calculator for Diagnostic Tests}
\usage{
sample_non_inferior_sesp(
  theta_S,
  theta_E,
  p_ts_pos_given_te_pos = NULL,
  delta,
  alpha = 0.05,
  beta,
  R = 1,
  paired = TRUE
)
}
\arguments{
\item{theta_S}{(numeric): The true sensitivity (or specificity) of an established standard test (test S).}

\item{theta_E}{(numeric): The true sensitivity (or specificity) of a new experimental test (test E).}

\item{p_ts_pos_given_te_pos}{(numeric): Conditional probability of the standard test being positive when experimental test being positive.}

\item{delta}{(numeric): The smallest difference in accuracy which would not be considered non-inferior.}

\item{alpha}{(numeric): Significance level (default: 0.05).}

\item{beta}{(numeric): Type II error rate.}

\item{R}{(numeric): Ratio of patients without to with the condition (default: 1).}

\item{paired}{(logical): Whether the design is paired (\code{TRUE}) or unpaired (\code{FALSE}).}
}
\value{
An object of class "diag_sample_size_html" containing:
\itemize{
\item \code{sample_size} (list): Required sample size.
\item \code{parameters} (list): Input parameters.
\item \code{html_report} (html): Sample size calculation report.
}
}
\description{
Calculate the required sample size for testing whether the sensitivity/specificity of a new experimental test (test E) is non-inferior to an established standard test (test E).
}
\examples{
n <- sample_non_inferior_sesp(theta_S=0.9, theta_E=0.89, p_ts_pos_given_te_pos=0.5, delta=0.1, alpha=0.05, beta=0.2, R=1, paired=TRUE)
print(n)
}
