% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_non_inferiority_auc.R
\name{sample_non_inferiority_auc}
\alias{sample_non_inferiority_auc}
\title{Sample Size Calculator for Diagnostic Tests}
\usage{
sample_non_inferiority_auc(
  A1,
  b1 = 1,
  A2,
  b2 = 1,
  r = NULL,
  rN = NULL,
  rD = NULL,
  delta,
  alpha = 0.05,
  beta,
  R = 1,
  paired = TRUE,
  dist = "any"
)
}
\arguments{
\item{r}{(numeric): The correlation between the tests because of the paired design}

\item{delta}{(numeric): The smallest unacceptable difference in accuracy}

\item{alpha}{(numeric): Significance level (default: 0.05)}

\item{beta}{(numeric): Type II error rate}

\item{R}{(numeric): Ratio of patients with and without the condition}

\item{paired}{(character): Paired study design or unpaired study design, must be "True" or "False"}

\item{A1_alter}{(numeric): The expected auc of test1}

\item{A2_alter}{(numeric): The expected auc of test2}
}
\value{
An object of class "diag_sample_size_html" containing:
\itemize{
\item \code{sample_size} (list): Required sample size.
\item \code{parameters} (list): Input parameters.
\item \code{html_report} (html): Sample size calculation report.
}
}
\description{
Calculate the required sample size for testing whether the Area Under the ROC Curve (AUC) of a new test (test1) is non-inferior to an existing test (test2).
}
\examples{
n <- sample_non_inferiority_auc(A1=0.9, A2=0.88, rN=0.5, rD=0.5, delta=0.05, alpha=0.05, beta=0.2, R=1, paired=TRUE, dist="binorm")
print(n)
}
