% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_hypotest_auc.R
\name{sample_hypotest_auc}
\alias{sample_hypotest_auc}
\title{Sample Size Calculation for Fixed AUC Comparison}
\usage{
sample_hypotest_auc(
  A_null,
  A_alter,
  b = 1,
  alpha = 0.05,
  beta,
  alternative = "two.sided",
  R = 1,
  dist = "any"
)
}
\arguments{
\item{A_null}{(numeric): AUC under the null hypothesis (default: NULL).}

\item{A_alter}{(numeric): AUC under the alternative hypothesis (default: NULL).}

\item{b}{(numeric): Ratio of the standard deviations of the distributions of results for patients without versus with the condition, σ_without/σ_with (default: 1).}

\item{alpha}{(numeric): Significance level (0 < alpha < 1).}

\item{beta}{(numeric): Type II error (0 < beta < 1, power = 1 - beta).}

\item{alternative}{(character): Type of test, must be one of "less", "greater", or "two.sided" (default: "two.sided").}

\item{R}{(numeric): Ratio of patients with and without the condition (default: 1).}

\item{dist}{(character): The assumption for choosing the variance function (default: "any"):
\itemize{
\item \code{"any"}: Applicable for all distributions.
\item \code{"exp"}: Assume that the unobserved, underlying test results follow an exponential distribution, but the observed test results, either continuous or ordinal, do not necessarily have an exponential distribution
\item \code{"binorm"}: Assume that the unobserved, underlying test results follow a binormal distribution, but the observed test results, either continuous or ordinal, do not necessarily have a binormal distribution
\item \code{"obs_binorm"}: Assume that the observed test results are on a truly continuous scale and follow a binormal distribution or can be transformed to a binormal distribution.
}}
}
\value{
An object of class "diag_sample_size_html" containing:
\itemize{
\item \code{sample_size} (list): Required sample size.
\item \code{parameters} (list): Input parameters.
\item \code{html_report} (html): Sample size calculation report.
}
}
\description{
Computes the required sample size for testing whether the AUC is equal to a specific value.
}
\details{
Sample Size Calculator for Diagnostic Tests

Calculate the required sample size for testing the hypothesis that the Area Under the ROC Curve of a new test is equal to a particular value.
}
\examples{
n <- sample_hypotest_auc(A_null=0.5, A_alter=0.8, alpha=0.05, beta=0.2, alternative="two.sided", R=1, dist="binorm")
print(n)
}
