% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_non_inferior_sesp.R
\name{sample_non_inferior_sesp}
\alias{sample_non_inferior_sesp}
\title{Sample Size Calculator for Diagnostic Tests}
\usage{
sample_non_inferior_sesp(
  theta1,
  theta2,
  prob_t1pos_give_t2pos = NULL,
  delta,
  alpha = 0.05,
  beta,
  R = 1,
  paired = TRUE
)
}
\arguments{
\item{theta1}{(numeric): The expected sensitivity (or specificity) of a new test (test 1).}

\item{theta2}{(numeric): The expected sensitivity (or specificity) of an existing test (test 2).}

\item{prob_t1pos_give_t2pos}{(numeric): Conditional probability of test 1 being positive when test 2 being positive.}

\item{delta}{(numeric): The smallest difference in accuracy which is not acceptable.}

\item{alpha}{(numeric): Significance level (default: 0.05).}

\item{beta}{(numeric): Type II error rate.}

\item{paired}{(logical): Whether the design is paired (\code{TRUE}) or unpaired (\code{FALSE}).}
}
\value{
An object of class "diag_sample_size_html" containing:
\itemize{
\item \code{sample_size} (list): Required sample size.
\item \code{parameters} (list): Input parameters.
\item \code{html_report} (html): Sample size calculation report.
}
}
\description{
Calculate the required sample size for testing whether the sensitivity/specificity of a new test (test1) is non-inferior to an existing test (test2).
}
\examples{
n <- sample_non_inferior_sesp(theta1=0.8, theta2=0.9, prob_t1pos_give_t2pos=0, delta=0.1, alpha=0.05, beta=0.2, R=1, paired=TRUE)
print(n)
}
